pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "nimbus"
version = "1.0.0"
description = "Dynamic Island clone for Windows — physics-animated, real system integrations"
readme = "README.md"
license = { text = "MIT" }
authors = [{ name = "nimbus Contributors" }]
requires-python = ">=3.10"
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Win32 (MS Windows)",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Operating System :: Microsoft :: Windows",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Desktop Environment",
]

dependencies = [
    "PySide6>=6.6.0",
    "psutil>=5.9.0",
    "Pillow>=10.0.0",
    "numpy>=1.24.0",
    "pywin32>=305;platform_system=='Windows'",
    "winrt-runtime;platform_system=='Windows'",
    "winrt-Windows.Media.Control;platform_system=='Windows'",
    "winrt-Windows.Foundation;platform_system=='Windows'",
    "winrt-Windows.Security.Authorization;platform_system=='Windows'",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-qt>=4.2.0",
    "ruff>=0.1.0",
    "mypy>=1.5.0",
    "types-pillow>=10.0.0",
]

[project.scripts]
nimbus = "nimbus.__main__:main"

[tool.hatch.build.targets.wheel]
packages = ["nimbus"]

[tool.ruff]
line-length = 100
target-version = "py310"
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "N",    # pep8-naming
    "UP",   # pyupgrade
    "B",    # flake8-bugbear
    "C4",   # flake8-comprehensions
    "T10",  # flake8-debugger
    "PIE",  # flake8-pie
    "RUF",  # ruff-specific
]
ignore = [
    "E501",  # line-too-long (handled by ruff format)
]

[tool.ruff.per-file-ignores]
"tests/**" = ["F811"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
strict_optional = true
check_untyped_defs = true
disallow_untyped_defs = true
disallow_any_unimported = false
disallow_any_expr = false
disallow_incomplete_defs = true
disallow_untyped_calls = false

[[tool.mypy.overrides]]
module = "win32com.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "psutil.*"
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
qt_api = "pyside6"
addopts = "-vvs"
